Move rateLimit into infrastructure/
git mv lib/rateLimit.ts -> infrastructure/rateLimit.ts and repoint the two
importers (booking + marketing mutations). Rounds out the target tree; the
remaining lib/ helpers (google, calendars, eventReads, assistantLogic,
googleCredentials, notifications) are genuinely shared leaves that stay put. No
new errors module — the provider error taxonomy already lives, correctly, in
integrations/calendar/errors.ts. Pure move; types + full suite green.
