+++ packages/backend/convex/_generated/api.d.ts
@@ -48,6 +48,7 @@ import type * as domains_sync_engine from "../domains/sync/engine.js";
 import type * as googleSync from "../googleSync.js";
 import type * as healthCheck from "../healthCheck.js";
 import type * as http from "../http.js";
+import type * as infrastructure_rateLimit from "../infrastructure/rateLimit.js";
 import type * as integrations_calendar_errors from "../integrations/calendar/errors.js";
 import type * as integrations_calendar_registry from "../integrations/calendar/registry.js";
 import type * as integrations_calendar_service from "../integrations/calendar/service.js";
@@ -61,7 +62,6 @@ import type * as lib_eventReads from "../lib/eventReads.js";
 import type * as lib_google from "../lib/google.js";
 import type * as lib_googleCredentials from "../lib/googleCredentials.js";
 import type * as lib_notifications from "../lib/notifications.js";
-import type * as lib_rateLimit from "../lib/rateLimit.js";
 import type * as maintenance from "../maintenance.js";
 import type * as migrations_backfills from "../migrations/backfills.js";
 import type * as notifications from "../notifications.js";
@@ -116,6 +116,7 @@ declare const fullApi: ApiFromModules<{
   googleSync: typeof googleSync;
   healthCheck: typeof healthCheck;
   http: typeof http;
+  "infrastructure/rateLimit": typeof infrastructure_rateLimit;
   "integrations/calendar/errors": typeof integrations_calendar_errors;
   "integrations/calendar/registry": typeof integrations_calendar_registry;
   "integrations/calendar/service": typeof integrations_calendar_service;
@@ -129,7 +130,6 @@ declare const fullApi: ApiFromModules<{
   "lib/google": typeof lib_google;
   "lib/googleCredentials": typeof lib_googleCredentials;
   "lib/notifications": typeof lib_notifications;
-  "lib/rateLimit": typeof lib_rateLimit;
   maintenance: typeof maintenance;
   "migrations/backfills": typeof migrations_backfills;
   notifications: typeof notifications;
